+++ src/main/java/com/solera/airline/controller/UserController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -30,8 +31,13 @@ public class UserController {
 	}
 	
 	@GetMapping("/{id}")
-	public User getUserById(@RequestParam int id) {
+	public User getUserById(@PathVariable int id) {
 		return uService.findByIdUser(id);
+	}
+	
+	@GetMapping("/name/{name}")
+	public User getUserByName(@PathVariable String name) {
+		return uService.findByNameUser(name);
 	}
 	
 	@PostMapping("/create")
@@ -39,8 +45,11 @@ public class UserController {
 		return uService.addUser(user) == 1 ? "ok" : "error";
 	}
 	
-	@DeleteMapping("/delete/{id}")
-	public String deleteUser(@RequestParam int id) {
+	@DeleteMapping("/{id}/delete")
+	public String deleteUser(@PathVariable int id) {
+		if(uService.findByIdUser(id) == null) {
+			return "The user does not exist";
+		}
 		return uService.deleteUser(id) == 1 ? "ok" : "error";
 	}
